requirements.txt
asgiref==3.7.1
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
drf-yasg==1.21.5
PyJWT==1.7.1
django-oauth-toolkit~=1.3.3
django-filter==2.4.0
pytest==7.3.1
pytest-django==3.7.0
six==1.15.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0



